+++ backend/package.json
@@ -33,6 +33,8 @@
     "@nestjs/websockets": "^9.3.2",
     "@prisma/client": "^4.10.1",
     "axios": "^1.3.4",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.0",
     "cookie-parser": "^1.4.6",
     "otplib": "^12.0.1",
     "passport": "^0.6.0",
