+++ backend/src/auth/auth.service.ts
@@ -12,8 +12,7 @@ import { authenticator } from 'otplib';
 import { toFileStream } from 'qrcode';
 import { Response } from 'express';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { SignUpUserDto } from 'src/user/dto/user.dto';
-import { AuthDto } from './dto/auth.dto';
+import { AuthDto, SignUpUserDto } from './dto/auth.dto';
 import { Jwt } from './type/auth.type';
 import { randomBytes, scrypt } from 'crypto';
 import { promisify } from 'util';
