+++ backend/src/auth/dto/auth.dto.ts
@@ -1,3 +1,4 @@
+import { ApiProperty } from '@nestjs/swagger';
 export class Msg {
   message: string;
 }
@@ -7,4 +8,17 @@ export class AuthDto {
 }
 export class OtpCodeDao {
   otpcode: string;
+}
+
+export class SignUpUserDto {
+  @ApiProperty()
+  name: string;
+  @ApiProperty()
+  email: string;
+  @ApiProperty()
+  password?: string;
+  @ApiProperty()
+  image: string;
+  @ApiProperty()
+  isFtLogin?: boolean;
 }
