--- backend/src/user/dto/user.dto.ts
@@ -1,14 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class SignUpUserDto {
-  @ApiProperty()
-  name: string;
-  @ApiProperty()
-  email: string;
-  @ApiProperty()
-  password?: string;
-  @ApiProperty()
-  image: string;
-  @ApiProperty()
-  isFtLogin?: boolean;
-}
